# Conflicts:
#	modules/Admin/Inventory/UpdateInventory.php

+++ controllers/EmployeeController.php
@@ -59,6 +59,12 @@ class EmployeeController
     public function add($data, $roleIds = [], $menuIds = [])
     {
         try {
+            if ($this->employeeModel->existsByName($data['name'])) {
+                return [
+                    'success' => false,
+                    'message' => 'Tên nhân viên đã tồn tại!'
+                ];
+            }
             $employeeId = $this->employeeModel->insert($data);
             foreach ($roleIds as $rid) {
                 $this->roleEmployeeModel->insert([
@@ -85,6 +91,23 @@ class EmployeeController
     public function update($id, $data, $roleIds = [], $menuIds = [])
     {
         try {
+
+            $existingById = $this->employeeModel->find($id);
+            if ($existingById == null) {
+                return [
+                    'success' => false,
+                    'message' => 'Nhân viên không tồn tại!'
+                ];
+            }
+
+            $existingByName = $this->employeeModel->existsByNameExceptId($id, $data['name']);
+            if ($existingByName) {
+                return [
+                    'success' => false,
+                    'message' => 'Tên nhân viên này đã tồn tại, vui lòng chọn tên khác.'
+                ];
+            }
+
             $this->employeeModel->update($id, $data);
 
             $this->roleEmployeeModel->deleteByEmployee(employeeId: $id);
@@ -113,6 +136,13 @@ class EmployeeController
 
     public function delete($id)
     {
+        $existingById = $this->employeeModel->find($id);
+        if ($existingById == null) {
+            return [
+                'success' => false,
+                'message' => 'Nhân viên không tồn tại!'
+            ];
+        }
         $this->employeeModel->updateDeleted($id);
         $this->roleEmployeeModel->deleteByEmployee(employeeId: $id);
         $this->employeeMenuModel->deleteByEmployee($id);
